+++ R/authors-at-r.R
@@ -3,7 +3,7 @@
 
 parse_authors_at_r <- function(x) {
 
-  if (is.null(x)) return(NULL)
+  if (is.null(x) || is.na(x)) return(NULL)
 
   out <- tryCatch(
     eval(parse(text = x)),
@@ -53,12 +53,14 @@ search_for_author <- function(authors, given = NULL, family = NULL,
 
 
 idesc_get_authors <- function(self, private, ensure = TRUE) {
+  assert_that(is_flag(ensure))
   if (ensure) ensure_authors_at_r(self)
   parse_authors_at_r(self$get("Authors@R"))
 }
 
 
 idesc_get_author <- function(self, private, role) {
+  assert_that(is_string(role))
   if (self$has_fields("Authors@R")) {
     aut <- self$get_authors()
     roles <- aut$role
@@ -72,23 +74,40 @@ idesc_get_author <- function(self, private, role) {
 }
 
 idesc_set_authors <- function(self, private, authors) {
+  assert_that(is_authors(authors))
   self$set("Authors@R", deparse_authors_at_r(authors))
 }
 
+check_author_args <- function(given = NULL, family = NULL, email = NULL,
+                              role = NULL, comment = NULL) {
+  assert_that(
+    is_string_or_null(given),
+    is_string_or_null(family),
+    is_string_or_null(email),
+    is_string_or_null(role),
+    is_string_or_null(comment)
+  )
+}
 
 #' @importFrom utils person
 
 idesc_add_author <- function(self, private, given, family, email, role,
-                            comment) {
+                             comment) {
+  check_author_args(given, family, email, role, comment)
   orig <- idesc_get_authors(self, private, ensure = FALSE)
   newp <- person(given = given, family = family, email = email,
                  role = role, comment = comment)
-  self$set_authors(c(orig, newp))
+  new_authors <- if (is.null(orig)) newp else c(orig, newp)
+  self$set_authors(new_authors)
 }
 
 
 idesc_add_role <- function(self, private, role, given, family, email,
-                          comment) {
+                           comment) {
+
+  assert_that(is.character(role))
+  check_author_args(given, family, email, comment = comment)
+
   orig <- idesc_get_authors(self, private, ensure = FALSE)
   wh <- search_for_author(
     orig, given = given, family = family, email = email, comment = comment,
@@ -110,6 +129,8 @@ idesc_add_role <- function(self, private, role, given, family, email,
 
 idesc_del_author <- function(self, private, given, family, email, role,
                             comment) {
+
+  check_author_args(given, family, email, role, comment)
 
   orig <- idesc_get_authors(self, private, ensure = FALSE)
   wh <- search_for_author(
@@ -135,6 +156,9 @@ idesc_del_author <- function(self, private, given, family, email, role,
 idesc_del_role <- function(self, private, role, given, family, email,
                           comment) {
 
+  assert_that(is.character(role))
+  check_author_args(given, family, email, role = NULL, comment)
+
   orig <- idesc_get_authors(self, private, ensure = FALSE)
   wh <- search_for_author(
     orig, given = given, family = family, email = email, comment = comment,
@@ -156,6 +180,7 @@ idesc_del_role <- function(self, private, role, given, family, email,
 
 idesc_change_maintainer <- function(self, private, given, family, email,
                                    comment) {
+  check_author_args(given, family, email, role = NULL, comment)
   ensure_authors_at_r(self)
   self$del_role(role = "cre")
   self$add_role(role = "cre", given = given, family = family,
@@ -166,6 +191,8 @@ idesc_change_maintainer <- function(self, private, given, family, email,
 #' @importFrom utils tail
 
 idesc_add_me <- function(self, private, role, comment) {
+  assert_that(is_string_or_null(role))
+  assert_that(is_string_or_null(comment))
   check_for_package("whoami", "$add_me needs the 'whoami' package")
   fn <- strsplit(whoami::fullname(), "[ ]+")[[1]]
   family <- tail(fn, 1)