+++ tests/testthat/test-deps.R
@@ -119,3 +119,42 @@ test_that("deleting a non-dependency is OK", {
   after <- desc$get("Imports")
   expect_equal(before, after)
 })
+
+test_that("has_dep", {
+
+  desc <- description$new("D1")
+  expect_true(desc$has_dep("R6"))
+  expect_true(desc$has_dep("testthat"))
+
+  expect_true(desc$has_dep("R6", "Imports"))
+  expect_true(desc$has_dep("testthat", "Suggests"))
+
+  expect_false(desc$has_dep("foobar"))
+
+  expect_false(desc$has_dep("testthat", "Imports"))
+
+  expect_error(desc$has_dep(123))
+  expect_error(desc$has_dep("testthat", "xxx"))
+})
+
+test_that("issue #34 is fine (empty dep fields)", {
+
+  empty_deps <- data.frame(
+    stringsAsFactors = FALSE,
+    type = character(),
+    package = character(),
+    version = character()
+  )
+
+  desc <- description$new("D4")
+  expect_silent(deps <- desc$get_deps())
+  expect_equal(deps, empty_deps)
+
+  desc$set(Imports = "")
+  expect_silent(deps <- desc$get_deps())
+  expect_equal(deps, empty_deps)
+
+  expect_silent(desc$set_deps(deps))
+  expect_silent(deps <- desc$get_deps())
+  expect_null(deps)
+})