+++ inst/NEWS.md
@@ -1,4 +1,13 @@
 
+# 1.1.0
+
+* Fix bug when adding authors and there is no `Authors@R` field
+
+* Get `DESCRIPTION` from package archives (#40)
+
+* Fix but in `del_dep()` and `has_dep()`, they only worked if the package
+  was attached.
+
 # 1.0.1
 
 * Fix formatting of `Collate` fields, they always start at a new line now.